requirements.txt
bcrypt~=4.0.1
fastapi~=0.88.0
PyJWT~=1.7.1
pydantic~=1.10.2
requests~=2.27.1
SQLAlchemy~=1.4.44
setuptools~=56.2.0
uvicorn~=0.20.0
python-dotenv~=0.19.2
python-decouple==3.6
